requirements.txt
Flask==1.1.1
magenta==1.1.7
tensorflow==1.15.0
urllib3==1.24.2
beautifulsoup4==4.8.0
pyFluidSynth==1.2.5
pyhumps==1.3.1